+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## FUTURE
+## 0.6.0
 
 * [ADDED] - new `--genesis-version` option to the `migrate` command
 * [ADDED] - `mig:status` command prints info about which verions are about to be migrated